Tidy test resources
