+++ tests/resources/klargjorte_data/person_testdata_p2021-12-31_p2021-12-31_v1.parquet
