+++ tests/utils.py
@@ -7,11 +7,12 @@ TEST_BUCKET_PARQUET_FILEPATH = "gs://ssb-staging-dapla-felles-data-delt/datadoc/
 TEST_BUCKET_PARQUET_FILEPATH_WITH_SHORTNAME = "gs://ssb-staging-dapla-felles-data-delt/befolkning/klargjorte_data/person_data_v1.parquet"
 
 TEST_RESOURCES_DIRECTORY = Path("tests/resources")
+TEST_DATASETS_DIRECTORY = TEST_RESOURCES_DIRECTORY / "datasets"
 
 TEST_PARQUET_FILE_NAME = "person_data_v1.parquet"
-TEST_PARQUET_FILEPATH = TEST_RESOURCES_DIRECTORY / TEST_PARQUET_FILE_NAME
-TEST_SAS7BDAT_FILEPATH = TEST_RESOURCES_DIRECTORY / "sasdata.sas7bdat"
-TEST_PARQUET_GZIP_FILEPATH = TEST_RESOURCES_DIRECTORY / "person_data_v1.parquet.gzip"
+TEST_PARQUET_FILEPATH = TEST_DATASETS_DIRECTORY / TEST_PARQUET_FILE_NAME
+TEST_SAS7BDAT_FILEPATH = TEST_DATASETS_DIRECTORY / "sasdata.sas7bdat"
+TEST_PARQUET_GZIP_FILEPATH = TEST_DATASETS_DIRECTORY / "person_data_v1.parquet.gzip"
 
 TEST_EXISTING_METADATA_DIRECTORY = TEST_RESOURCES_DIRECTORY / "existing_metadata_file"
 TEST_EXISTING_METADATA_FILE_NAME = "person_data_v1__DOC.json"
@@ -25,15 +26,4 @@ TEST_EXISTING_METADATA_WITH_VALID_ID_DIRECTORY = (
 
 TEST_COMPATIBILITY_DIRECTORY = TEST_EXISTING_METADATA_DIRECTORY / "compatibility"
 
-TEST_INPUT_DATA_POPULATION_DIRECTORY = (
-    TEST_RESOURCES_DIRECTORY / "inndata" / "befolkning"
-)
-TEST_PROCESSED_DATA_POPULATION_DIRECTORY = (
-    TEST_RESOURCES_DIRECTORY / "klargjorte_data" / "befolkning"
-)
-TEST_STATISTICS_POPULATION_DIRECTORY = (
-    TEST_RESOURCES_DIRECTORY / "statistikk" / "befolkning"
-)
-TEST_OUTPUT_DATA_POPULATION_DIRECTORY = (
-    TEST_RESOURCES_DIRECTORY / "utdata" / "befolkning"
-)
+TEST_PROCESSED_DATA_POPULATION_DIRECTORY = TEST_RESOURCES_DIRECTORY / "klargjorte_data"
